% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TFs.R
\name{get_TFs}
\alias{get_TFs}
\title{Retrieve Transcription Factors Regulating a Given Target Gene}
\usage{
get_TFs(target_gene, validation_freq = 1, max_size = NULL)
}
\arguments{
\item{target_gene}{A string specifying the name of the Target Gene.}

\item{validation_freq}{A numeric value representing the validation frequency threshold. Default is 1.}

\item{max_size}{An integer that limits the maximum size of the returned vector of TFs. Default is NULL, meaning no limit.}
}
\value{
A vector of TFs that regulate the specified Target Gene, limited in size by max_size.
}
\description{
This function returns a vector of Transcription Factors (TFs) that regulate a specified Target Gene.
It filters the data based on the Target Gene and a validation frequency, and limits the size of the returned vector.
}
\examples{
get_TFs("Gene1")
get_TFs("Gene2", validation_freq = 2)
get_TFs("Gene3", max_size = 5)

}
