% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TF_regulon.R
\name{plot_TF_regulon}
\alias{plot_TF_regulon}
\title{Plot Regulation Graph for a Gene with Size Limits}
\usage{
plot_TF_regulon(gene_name, max_TFs = NULL, max_targets = NULL)
}
\arguments{
\item{gene_name}{A string specifying the name of the gene.}

\item{max_TFs}{An integer limiting the maximum number of TFs shown. Default is NULL, meaning no limit.}

\item{max_targets}{An integer limiting the maximum number of target genes shown. Default is NULL, meaning no limit.}
}
\description{
This function plots a graph showing all the transcription factors (TFs) that regulate a given gene
and all the target genes regulated by it, with optional limits on the number of TFs and targets shown.
}
\examples{
plot_TF_regulon("CTCF")
plot_TF_regulon("CTCF", max_TFs = 10, max_targets = 10)

}
