% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TF_TG_val.R
\name{plot_TF_TG_val}
\alias{plot_TF_TG_val}
\title{Plot Target Gene Validation Scores for a Given TF with Frequency and Size Filters}
\usage{
plot_TF_TG_val(tf_name, validation_freq = 1, max_size = 30)
}
\arguments{
\item{tf_name}{A string specifying the name of the Transcription Factor.}

\item{validation_freq}{A numeric value representing the validation frequency threshold. Default is 1.}

\item{max_size}{An integer that limits the maximum number of target genes for which scores are plotted. Default is NULL, meaning no limit.}
}
\value{
A value of 1 indicating that the plot function has run without any errors
}
\description{
This function plots the target genes against their TF_TG_Exp_qual_score for a specified Transcription Factor (TF),
considering the frequency of the count and limiting the number of target genes based on max_size.
Each target gene is represented by a different color to form a spectrum.
The TF_TG_Exp_qual_score in the TF_TG_Valid_Comb datasets is a qualitative
score for each TF-TG interaction based on validation across datasets and in
comparison to other TF-TG's in the combined datasets.
}
\examples{

# Plot the validation score of the target genes of the input TF.
plot_TF_TG_val("CTCF")

# Allow the user to specify the validation frequency of the target genes
# that are to be displayed, a number between 1 and 4. Also allows the user
# to specify the number of target genes to show using max_size. The default
# value is 30 to optimize the graph visual.
plot_TF_TG_val("CTCF", validation_freq = 2, max_size = 10)


}
