% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TF_Total_Val.R
\name{plot_TF_total_val}
\alias{plot_TF_total_val}
\title{Plot Validation Scores for Multiple Transcription Factors}
\usage{
plot_TF_total_val(tf_names)
}
\arguments{
\item{tf_names}{A vector of strings specifying the names of the Transcription Factors.}
}
\value{
The plot graph.
}
\description{
This function takes a vector of Transcription Factor (TF) names and plots a
graph of their names against their Validation Scores. The validation score
is represented by the Ratio_to_Overall_Avg from the TF_Avg_Valid dataset
which is a qualitative value that represents the normalized average
validation quality of a specific TF over all TF-TG interactions in the
combined dataset that involved that TF.
}
\examples{
# Plots the the normalized average validation score for an input array of TFs.
plot_TF_total_val(c("CTCF", "GATA1", "AHR", "AHRR", "BDP1"))

}
