% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TFs_network.R
\name{plot_TFs_network}
\alias{plot_TFs_network}
\title{Plot Combined Regulation Graph for Up to Three Genes}
\usage{
plot_TFs_network(gene_names, max_TFs = NULL, max_targets = NULL)
}
\arguments{
\item{gene_names}{A vector of up to three gene names.}

\item{max_TFs}{An integer limiting the maximum number of TFs shown for each gene. Default is NULL, meaning no limit.}

\item{max_targets}{An integer limiting the maximum number of target genes shown for each gene. Default is NULL, meaning no limit.}
}
\description{
This function plots a combined graph showing all the transcription factors (TFs) that regulate up to three given genes
and all the target genes regulated by them. The input genes, their regulators, and their target genes are colored differently.
}
\examples{
plot_TFs_network(c("CTCF", "GATA1"))
plot_TFs_network(c("CTCF", "GATA1", "STAT1"), max_TFs = 10, max_targets = 10)

}
