% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTRNValStandVis.R
\name{runTRNValStandVis}
\alias{runTRNValStandVis}
\title{Launch Shiny App for TRNValStandVis}
\usage{
runTRNValStandVis()
}
\value{
No return value but opens up a Shiny page.
}
\description{
A function that launches the Shiny app for the TRNValStandVis package.
This app visualizes, validates, and standardizes TF-TG interactions
and Transcription Regulatory Networks (TRN). For more information
on the package, refer to its GitHub repository.
}
\examples{
\dontrun{

TRNValStandVis::runTRNValStandVis()
}

}
