% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TF_Avg_Valid}
\alias{TF_Avg_Valid}
\title{Dataset of the overall validation score of individual Transcription Factors.}
\format{
A dataframe with 2 columns:
\describe{
\item{TF}{Transcription Factor name in HGNC. }
\item{Ratio_to_Overall_Avg}{Average validation score for all TFs interactions.}

...
}
}
\usage{
data(TF_Avg_Valid)
}
\description{
This dataset maps every HGNC standardized TF to its overal avg validation score
over all it TF-TG interactions
}
\examples{
\dontrun{
 TF_Avg_Valid
}

}
\keyword{datasets}
