% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TF_TG_Valid_Comb}
\alias{TF_TG_Valid_Comb}
\title{Dataset of all the TF-TG interactions combined from the four sources.}
\format{
A data frame with 5 columns:
\describe{
\item{TF}{Transcription Factor name in HGNC. }
\item{Target_Gene}{Target Gene name in HGNC.}
\item{Count}{Frequency of TF-TG across the datasets.}
\item{TF_TG_Exp_qual_score}{Exponential validation score of TF-TG based on frequency}
\item{Ratio_to_TF_Avg}{Normalized validation score based on TF_TG_Exp_qual_score.}

...
}
}
\usage{
data(TF_TG_Valid_Comb)
}
\description{
This dataset combines data from the ChEA3, hTFtarget, MSigDB and TFLink to
create a dataset of all the TF-TG interactions available in all the 4 datasets.
To process the data, all genes that did not have a translatable HGNC symbol,
even in alternate nomenclatures was removed. Based on the count of appearance
of a TF-TG in the datasets a validation score was created.
TF_TG_Exp_qual_score is a qualitative score for each TF-TG interaction based
on validation across datasets and in comparison to other TF-TG's in the
combined datasets. Ratio_to_TF_Avg is a normalized validation score based on
each TF_TG_Exp_qual_score relative to that for all the the scores for that TF.
}
\details{
Description of the dataset.
}
\examples{
\dontrun{
 TF_TG_Valid_Comb
}

}
\keyword{datasets}
