% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TF_TG_val.R
\name{get_TF_TG_val}
\alias{get_TF_TG_val}
\title{Retrieve validation score for all Target Genes(TG) of a given Transcription Factor(TF)}
\usage{
get_TF_TG_val(tf_name, validation_freq = 1, max_size = NULL)
}
\arguments{
\item{tf_name}{A string specifying the name of the Transcription Factor.}

\item{validation_freq}{A numeric value representing the validation frequency threshold. Default is 1. Maximum is 4}

\item{max_size}{An integer that limits the maximum number of target genes for which scores are returned. Default is NULL, meaning no limit.}
}
\value{
A data frame with columns for target genes and their corresponding TF_TG_Exp_qual_score.
}
\description{
This function returns the validation score for all target genes of a specified franscription factor,
considering the frequency of the count, and limits the number of target genes based on max_size.
The validation score is represented by TF_TG_Exp_qual_score in the TF_TG_Valid_Comb datasets
and is a qualitative score for each TF-TG interaction based on validation across
datasets and in comparison to other TF-TG's in the combined datasets.
}
\examples{

# Gets all the target genes of the input gene and there validation score
get_TF_TG_val("CTCF")

# Gets the target genes of the input gene. Allows the user to specify
# the validation frequency between 1 and 4 and the max_size indicating the
# number of targets to show. If set to NULL will return all the targets
get_TF_TG_val("CTCF", validation_freq = 3, max_size = 10)

}
