% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TF_Total_Val.R
\name{get_TF_Total_Val}
\alias{get_TF_Total_Val}
\title{Retrieve Score for a Given Transcription Factor from tf_metrics}
\usage{
get_TF_Total_Val(tf_name)
}
\arguments{
\item{tf_name}{A string specifying the name of the Transcription Factor.}
}
\value{
The Ratio_to_Overall_Avg score for the specified TF.
}
\description{
This function returns the Ratio_to_Overall_Avg score for a specified Transcription Factor (TF)
from the TF_Avg_Valid dataset. The Ratio_to_Overall_Avg is qualitative value
that represents the normalized average validation quality of a specific
TF over all TF-TG interactions in the combined dataset that involved that
TF.
}
\examples{

# Gets the normalized average validation score for a specific TF.
get_TF_Total_Val("CTCF")

}
