% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TGs.R
\name{get_TGs}
\alias{get_TGs}
\title{Retrieve Target Genes for a Given Transcription Factor}
\usage{
get_TGs(tf_name, validation_freq = 1, max_size = NULL)
}
\arguments{
\item{tf_name}{A string specifying the name of the Transcription Factor.}

\item{validation_freq}{A numeric value representing the validation frequency threshold. Default is 1.}

\item{max_size}{An integer that limits the maximum size of the returned vector of Target Genes. Default is NULL, meaning no limit.}
}
\value{
A vector of Target Genes associated with the specified TF.
}
\description{
This function returns a vector of Target Genes associated with a specified Transcription Factor (TF) name.
It filters the data based on the TF name and a validation frequency.
}
\examples{
get_TGs("CTCF")
get_TGs("CTCF", validation_freq = 2)

}
