% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TF_regulon.R
\name{plot_TF_regulon}
\alias{plot_TF_regulon}
\title{Plot Regulation Graph for a Gene with Size Limits}
\usage{
plot_TF_regulon(gene_name, max_TFs = 30, max_targets = 30)
}
\arguments{
\item{gene_name}{A string specifying the name of the gene.}

\item{max_TFs}{An integer limiting the maximum number of TFs shown. Default is 30, for optimal visualization.}

\item{max_targets}{An integer limiting the maximum number of target genes shown. Default is 30, for optimal visualization.}
}
\value{
A value of 1 indicating that the plot function has run without any errors.

The output graph shows the input gene and it's target genes in orange

and it's regulating TF's in green. With arrows point from TF to TG.
}
\description{
This function plots a graph showing all the transcription factors (TFs) that regulate a given gene
and all the target genes regulated by it, with optional limits on the number of TFs and targets shown.
}
\examples{
# Plots the regulon of the gene CTCF showing it's regulators and targets
plot_TF_regulon("CTCF")

# Allows you to specify the maximum number of regulators and targets.
# The default is 30 genes for each to create the optimum output.
# There is no maximum and if the value set is greater than the available
# genes then all available genes are plotted.
plot_TF_regulon("CTCF", max_TFs = 10, max_targets = 10)

}
