% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TF_TG_val.R
\name{plot_TF_TG_val}
\alias{plot_TF_TG_val}
\title{Plot Target Gene Validation Scores for a Given TF with Frequency and Size Filters}
\usage{
plot_TF_TG_val(tf_name, validation_freq = 1, max_size = NULL)
}
\arguments{
\item{tf_name}{A string specifying the name of the Transcription Factor.}

\item{validation_freq}{A numeric value representing the validation frequency threshold. Default is 1.}

\item{max_size}{An integer that limits the maximum number of target genes for which scores are plotted. Default is NULL, meaning no limit.}
}
\description{
This function plots the target genes against their TF_TG_Exp_qual_score for a specified Transcription Factor (TF),
considering the frequency of the count and limiting the number of target genes based on max_size.
Each target gene is represented by a different color to form a spectrum.
}
\examples{
plot_TF_TG_val("CTCF")
plot_TF_TG_val("CTCF", validation_freq = 2, max_size = 10)


}
