% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TF_Total_Val.R
\name{plot_TF_total_val}
\alias{plot_TF_total_val}
\title{Plot Validation Scores for Multiple Transcription Factors}
\usage{
plot_TF_total_val(tf_names)
}
\arguments{
\item{tf_names}{A vector of strings specifying the names of the Transcription Factors.}
}
\description{
This function takes a vector of Transcription Factor (TF) names and plots a graph of their names against their Validation Scores.
}
\examples{
plot_TF_total_val(c("CTCF", "GATA1", "AHR", "AHRR", "BDP1"))
}
